Add hidden apply command for easy testing
